requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1
django-cotton==0.9.23
django-htmx==1.19.0
django-unfold==0.34.0
django-user-agents==0.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
ua-parser==0.18.0
user-agents==2.2.0
